Add Pancake imx ethernet driver

+++ include/sddf/network/queue.h
@@ -13,20 +13,23 @@
 #include <sddf/util/fence.h>
 #include <sddf/util/util.h>
 
+// Pancake Todo: Change `uint64_t len` etc back to same int type as C sDDF network queue
+// when pancake supports non-shared memory load/store for 16 and 32 bits
+
 typedef struct net_buff_desc {
     /* offset of buffer within buffer memory region or io address of buffer */
     uint64_t io_or_offset;
     /* length of data inside buffer */
-    uint16_t len;
+    uint64_t len;
 } net_buff_desc_t;
 
 typedef struct net_queue {
     /* index to insert at */
-    uint16_t tail;
+    uint64_t tail;
     /* index to remove from */
-    uint16_t head;
+    uint64_t head;
     /* flag to indicate whether consumer requires signalling */
-    uint32_t consumer_signalled;
+    uint64_t consumer_signalled;
     /* buffer descripter array */
     net_buff_desc_t buffers[];
 } net_queue_t;
@@ -37,7 +40,7 @@ typedef struct net_queue_handle {
     /* filled buffers */
     net_queue_t *active;
     /* size of the queues */
-    uint32_t size;
+    uint64_t size;
 } net_queue_handle_t;
 
 /**
@@ -200,7 +203,7 @@ static inline int net_dequeue_active(net_queue_handle_t *queue, net_buff_desc_t 
  * @param active pointer to active queue in shared memory.
  * @param size size of the free and active queues.
  */
-static inline void net_queue_init(net_queue_handle_t *queue, net_queue_t *free, net_queue_t *active, uint32_t size)
+static inline void net_queue_init(net_queue_handle_t *queue, net_queue_t *free, net_queue_t *active, uint64_t size)
 {
     queue->free = free;
     queue->active = active;
@@ -215,7 +218,7 @@ static inline void net_queue_init(net_queue_handle_t *queue, net_queue_t *free, 
  */
 static inline void net_buffers_init(net_queue_handle_t *queue, uintptr_t base_addr)
 {
-    for (uint32_t i = 0; i < queue->size - 1; i++) {
+    for (uint64_t i = 0; i < queue->size - 1; i++) {
         net_buff_desc_t buffer = {(NET_BUFFER_SIZE * i) + base_addr, 0};
         int err = net_enqueue_free(queue, buffer);
         assert(!err);
